debug

+++ src/e2eflow/core/flow_util.py
@@ -26,7 +26,7 @@ def flow_intensity(flow, threshold):
     mag = tf.expand_dims(mag, -1)
     shape = tf.shape(mag)
     condition = tf.greater(mag, threshold)
-    thresholded = tf.where(condition, tf.ones(shape, dtype=tf.int32), tf.zeros(shape, dtype=tf.int32))
+    thresholded = tf.where(condition, tf.ones(shape, dtype=tf.int16), tf.zeros(shape, dtype=tf.int16))
     return tf.multiply(thresholded, 255)
